requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
